speed up readOne and readLine by using io.ByteReader
and using an unbuffered byte reader wrapper if the argument
to Load()/NewUnpickler() doesn't support ReadByte().

This greatly reduces the allocations, since we're no longer
allocating a 1 byte buffer, reading into it, and throwing
it away to be gc'ed. My benchmarks exhibit a 47% reduction
in allocs, and a 18% reduction in CPU.

The wrapper has to be unbuffered because callers might call
NewUnpickler() repeatedly, passing in the same input stream
every time. The test cases in ./pytorch/ do this.

this speeds up by benchmarks by