slight speedup on memos by overallocating a little
go 1.19 seems to like to have a little spare room in a map,
at least one where the keys are going to be 0 to 255. Maybe
there are collisions in such a small set of keys.

Also preallocate a little room for Set and List, like we
do for Dict. In my benchmark the improvement is modest, but
then again my benchmarks don't have a lot of Set in them.